Create Linear regression and Random forest
In this project, I worked with a housing dataset to predict house prices based on housing variables. a Linear regression algorithm was used to train the model and to predict the prices. A random forest algorithm was also applied and both models were compared. feature importance was also done in other to identify the variables and their respective impact on the prediction.