Add files via upload
In this project, i worked on a housing dataset. I used a simple linear regression to create a model by first splitting the dataset into a training batch and a test batch, and then training the model with the training dataset and predicting the house prices with the test dataset. I also ran a Random forest algorithm and then compared both models to which gives a better result. The feature importance was also derived in other to see each variable's impact on the predicted prices.